+++ src/debugger/variable/mod.rs
@@ -10,6 +10,7 @@ pub mod dqe;
 pub mod execute;
 pub mod mutability;
 pub mod render;
+pub mod storage;
 pub mod value;
 pub mod r#virtual;